refactoring the Quiz Section App.js
