+++ src/Components/App.js
@@ -1,6 +1,4 @@
 //Main App
-
-import { useEffect, useReducer } from "react";
 import Header from "./Header";
 import Main from "./Main";
 import Loader from "./Loader";
@@ -13,109 +11,21 @@ import FinishedScreen from "./FinishedScreen";
 import Restart from "./Restart";
 import Timer from "./Timer";
 import Footer from "./Footer";
-
-const SECS_PER_QUESTION = 30;
-
-const initialState = {
-  questions: [],
-  status: "loading", //'loadng state' 'error state', 'ready state', 'active state', 'finished state'
-  index: 0,
-  answer: null,
-  points: 0,
-  highScore: 0,
-  secondsRemaining: null,
-};
-
-function reducer(state, action) {
-  switch (action.type) {
-    case "dataRecieved":
-      return {
-        ...state,
-        questions: action.payload,
-        status: "ready",
-      };
-
-    case "dataFailed":
-      return {
-        ...state,
-        status: "error",
-      };
-
-    case "start":
-      return {
-        ...state,
-        status: "active",
-        secondsRemaining: state.questions.length * SECS_PER_QUESTION,
-      };
-
-    case "newAnswer":
-      const questions = state.questions.at(state.index);
-      return {
-        ...state,
-        answer: action.payload,
-        points:
-          action.payload === questions.correctOption
-            ? state.points + questions.points
-            : state.points,
-      };
-
-    case "nextQuestion":
-      return {
-        ...state,
-        index: state.index + 1,
-        answer: null,
-      };
-
-    case "finished":
-      return {
-        ...state,
-        status: "finished",
-        highScore:
-          state.points > state.highScore ? state.points : state.highScore,
-      };
-
-    case "restart":
-      return {
-        ...initialState,
-        questions: state.questions,
-        status: "ready",
-      };
-
-    // return {
-    //   ...state,
-    //   points: 0,
-    //   highScore: 0,
-    //   index: 0,
-    //   answer: null,
-    //   state: "ready",
-    // };
-    case "timer":
-      return {
-        ...state,
-        secondsRemaining: state.secondsRemaining - 1,
-        status: state.secondsRemaining === 0 ? "finished" : state.status,
-      };
-
-    default:
-      throw new Error("Action is Unknown");
-  }
-}
+import { useQuiz } from "../Context/QuizProvider";
 
 export default function App() {
-  const [
-    { questions, status, index, answer, points, highScore, secondsRemaining },
+  const {
+    status,
+    numQuestioin,
     dispatch,
-  ] = useReducer(reducer, initialState);
-
-  const numQuestioin = questions.length;
-  const maxPoints = questions.reduce((prev, cur) => prev + cur.points, 0);
-
-  useEffect(function () {
-    fetch("http://localhost:8000/questions")
-      .then((res) => res.json())
-      .then((data) => dispatch({ type: "dataRecieved", payload: data }))
-      .catch((err) => dispatch({ type: "dataFailed" }));
-  }, []);
+    index,
+    points,
+    maxPoints,
+    answer,
+    highScore,
+    questions,
+    secondsRemaining,
+  } = useQuiz();
   return (
     <div className="app">
       <Header />
