+++ src/Components/Start.js
@@ -1,4 +1,3 @@
-import { type } from "@testing-library/user-event/dist/type";
 import React from "react";
 
 function Start({ numQuestioin, dispatch }) {
