+++ src/Context/QuizProvider.js
@@ -1,0 +1,135 @@
+import { createContext, useContext, useEffect, useReducer } from "react";
+
+const SECS_PER_QUESTION = 30;
+
+const QuizContext = createContext();
+
+const initialState = {
+  questions: [],
+  status: "loading",
+  index: 0,
+  answer: null,
+  points: 0,
+  highScore: 0,
+  secondsRemaining: null,
+};
+
+function reducer(state, action) {
+  switch (action.type) {
+    case "dataRecieved":
+      return {
+        ...state,
+        questions: action.payload,
+        status: "ready",
+      };
+
+    case "dataFailed":
+      return {
+        ...state,
+        status: "error",
+      };
+
+    case "start":
+      return {
+        ...state,
+        status: "active",
+        secondsRemaining: state.questions.length * SECS_PER_QUESTION,
+      };
+
+    case "newAnswer":
+      const questions = state.questions.at(state.index);
+      return {
+        ...state,
+        answer: action.payload,
+        points:
+          action.payload === questions.correctOption
+            ? state.points + questions.points
+            : state.points,
+      };
+
+    case "nextQuestion":
+      return {
+        ...state,
+        index: state.index + 1,
+        answer: null,
+      };
+
+    case "finished":
+      return {
+        ...state,
+        status: "finished",
+        highScore:
+          state.points > state.highScore ? state.points : state.highScore,
+      };
+
+    case "restart":
+      return {
+        ...initialState,
+        questions: state.questions,
+        status: "ready",
+      };
+
+    // return {
+    //   ...state,
+    //   points: 0,
+    //   highScore: 0,
+    //   index: 0,
+    //   answer: null,
+    //   state: "ready",
+    // };
+    case "timer":
+      return {
+        ...state,
+        secondsRemaining: state.secondsRemaining - 1,
+        status: state.secondsRemaining === 0 ? "finished" : state.status,
+      };
+
+    default:
+      throw new Error("Action is Unknown");
+  }
+}
+
+function QuizProvider({ children }) {
+  const [
+    { questions, status, index, answer, points, highScore, secondsRemaining },
+    dispatch,
+  ] = useReducer(reducer, initialState);
+
+  const numQuestioin = questions.length;
+  const maxPoints = questions.reduce((prev, cur) => prev + cur.points, 0);
+
+  useEffect(function () {
+    fetch("http://localhost:8000/questions")
+      .then((res) => res.json())
+      .then((data) => dispatch({ type: "dataRecieved", payload: data }))
+      .catch((err) => dispatch({ type: "dataFailed" }));
+  }, []);
+
+  return (
+    <QuizContext.Provider
+      value={{
+        questions,
+        status,
+        index,
+        answer,
+        points,
+        highScore,
+        secondsRemaining,
+        numQuestioin,
+        maxPoints,
+        dispatch,
+      }}
+    >
+      {children}
+    </QuizContext.Provider>
+  );
+}
+
+function useQuiz() {
+  const context = useContext(QuizContext);
+  if (context === undefined)
+    throw new Error("QuizContext used outside QuizProvider");
+  return context;
+}
+
+export { QuizProvider, useQuiz };
